아이템 1종(전자기 보호막) 기능 구현 및 아티메 스테이터스에 따른 프립펩 사용과 사용 구현

+++ Assets/Script/character/Enemy/EnemyController.cs
@@ -14,8 +14,8 @@ public class EnemyController : CharacterController
     {
         public StateEnum stateEnum; // 상태 열거형
         public float range; // 사거리, 99는 사거리 없음
-        public float cooldown; // 쿨타임, 0 쿨타임 없음
-        public float nowCoolTIme;//현제 쿨타임
+        public int cooldown; // 쿨타임, 0 쿨타임 없음
+        public int nowCoolTIme;//현제 쿨타임
         public bool NeedToPrepare;//파라미터
     }
